+++ rmf_task/src/rmf_task/Priority.cpp
@@ -1,0 +1,44 @@
+/*
+ * Copyright (C) 2021 Open Source Robotics Foundation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+*/
+
+#include "Priority.hpp"
+
+namespace rmf_task {
+
+//==============================================================================
+class BinaryPriority::Implementation
+{
+public:
+
+  std::size_t value;
+};
+
+//==============================================================================
+BinaryPriority::BinaryPriority(std::size_t value)
+: _pimpl(rmf_utils::make_impl<Implementation>(
+  Implementation{value}))
+{
+  // Do nothing
+}
+
+//==============================================================================
+std::size_t BinaryPriority::value() const
+{
+  return _pimpl->value;
+}
+
+}
